Add files via upload
It is all about Uncertainty Quantification of 2.5D Spectral Induced Polarization Inversion in a Bayesian Framework.
Our primary focus in these codes is to develop a 2.5D inversion code for full-waveform IP data using the Bayesian approach. Additionally, we aim to conduct an uncertainty analysis of the inverted model parameters, which is a crucial step in inversion, and provide a full stochastic analysis of the uncertainty in the IP inverse problem.